# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  handler.createOTCOfferRequest:
    properties:
      account_id:
        type: integer
      counterparty_system_type:
        type: string
      counterparty_user_id:
        type: integer
      direction:
        type: string
      on_behalf_of_client_id:
        type: integer
      quantity:
        type: string
      ticker:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: EXBanka Banking Microservices API Gateway. All endpoints are served
    under /api/v3 — v1 and v2 have been retired (plan E, 2026-04-27). Future versions
    (v4+) will be added as separate explicit router files with no transparent fallback.
    See api-gateway/internal/router/router_versioning.md.
  title: EXBanka API
  version: "3.0"
paths:
...
  /api/v2/me/holdings/{id}/transactions:
    get:
      description: Returns the OrderTransactions that contributed to a holding — when
        each buy/sell executed, price per unit, native vs account currency, FX rate,
        commission, and which account was used. Replaces the per-purchase detail removed
        from /me/portfolio in Part C.
      parameters:
      - description: Holding ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by direction (buy|sell); empty for both
        in: query
        name: direction
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: transactions + total_count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation_error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: not_found — holding does not exist or does not belong to caller
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List per-purchase history for a holding
      tags:
      - portfolio
...
  /api/v3/me/otc/options:
    get:
      description: Same response shape as GET /api/v3/otc/options (kind/bank_code/best_bid/best_ask/active_chains_count/...)
        but scoped to listings the caller posted. Only OPEN listings appear here (the
        unified cache is open-only); for a full history of every offer you've ever
        posted use GET /api/v3/me/otc/options/posted.
      parameters:
      - description: Filter to one ticker
        in: query
        name: ticker
        type: string
      - description: sell_initiated|buy_initiated
        in: query
        name: direction
        type: string
      - description: 1-based, default 1
        in: query
        name: page
        type: integer
      - description: default 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Marketplace view of the caller's OWN open OTC option listings
      tags:
      - OTCOptions
    post:
      consumes:
      - application/json
      description: Posts an open OTC option offer. Terms (strike_price, premium, settlement_date)
        are NOT set at creation — they are agreed during negotiation. Only one open
        offer per (owner, ticker, direction) is allowed; a duplicate returns 409.
      parameters:
      - description: offer details (ticker-keyed, open terms; account_id is the initiator's
          account)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createOTCOfferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an OTC option offer
      tags:
      - OTCOptions
...
  /api/v3/otc/options:
    get:
      description: Phase 6 cross-bank discovery. Returns this bank's open
      parameters:
      - description: Filter to one ticker
        in: query
        name: ticker
        type: string
      - description: local|remote
        in: query
        name: kind
        type: string
      - description: Filter to one bank
        in: query
        name: bank_code
        type: string
      - description: sell_initiated|buy_initiated
        in: query
        name: direction
        type: string
      - description: 1-based, default 1
        in: query
        name: page
        type: integer
      - description: default 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unified marketplace view of OPEN OTC option listings
      tags:
      - OTCOptions
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"